Move bubbles to modules and add JSDocs
